feat: TopBooksRankingコンポーネントを移管
